remove inactive revisions from status overview
