+++ k8s/statuser.go
@@ -5,11 +5,8 @@ import (
 	"fmt"
 	"github.com/nullstone-io/deployment-sdk/app"
 	"github.com/nullstone-io/deployment-sdk/logging"
-	v1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
-	"strconv"
-	"time"
 )
 
 type Statuser struct {
@@ -41,7 +38,12 @@ func (s Statuser) StatusOverview(ctx context.Context) (any, error) {
 		return so, fmt.Errorf("error retrieving app replica sets: %w", err)
 	}
 	for _, replicaSet := range replicaSets.Items {
-		so.Revisions = append(so.Revisions, RevisionFromReplicaSet(replicaSet))
+		revision := RevisionFromReplicaSet(replicaSet)
+		if revision.DesiredReplicas == 0 && revision.Replicas == 0 {
+			// Don't show old revisions that have scaled down
+			continue
+		}
+		so.Revisions = append(so.Revisions, revision)
 	}
 	return so, nil
 }
@@ -50,37 +52,6 @@ func (s Statuser) Status(ctx context.Context) (any, error) {
 	status := AppStatus{}
 
 	return status, nil
-}
-
-type AppStatusOverview struct {
-	Revisions []AppStatusOverviewRevision `json:"revisions"`
-}
-
-type AppStatusOverviewRevision struct {
-	Name              string    `json:"name"`
-	Revision          string    `json:"revision"`
-	Generation        int64     `json:"generation"`
-	CreatedAt         time.Time `json:"createdAt"`
-	DesiredReplicas   int       `json:"desiredReplicas"`
-	AvailableReplicas int       `json:"availableReplicas"`
-	ReadyReplicas     int       `json:"readyReplicas"`
-}
-
-func RevisionFromReplicaSet(rs v1.ReplicaSet) AppStatusOverviewRevision {
-	desired := 0
-	if val, err := strconv.Atoi(rs.Annotations["deployment.kubernetes.io/desired-replicas"]); err == nil {
-		desired = val
-	}
-
-	return AppStatusOverviewRevision{
-		Name:              rs.Name,
-		Revision:          rs.Annotations["deployment.kubernetes.io/revision"],
-		Generation:        rs.Status.ObservedGeneration,
-		CreatedAt:         rs.CreationTimestamp.Time,
-		DesiredReplicas:   desired,
-		AvailableReplicas: int(rs.Status.AvailableReplicas),
-		ReadyReplicas:     int(rs.Status.ReadyReplicas),
-	}
 }
 
 type AppStatus struct {
